+++ test/esri-loader.spec.js
@@ -32,6 +32,9 @@ describe('esri-loader', function () {
       it('should default to latest version', function () {
         expect(scriptEl.src).toEqual('https://js.arcgis.com/4.5/');
       });
+      it('should not have set dojoConfig', function () {
+        expect(window.dojoConfig).not.toBeDefined();
+      });
     });
     describe('with different API version', function () {
       var scriptEl;
@@ -55,6 +58,29 @@ describe('esri-loader', function () {
       });
       it('should have called callback', function () {
         expect(context.bootstrapCallback).toHaveBeenCalledTimes(1);
+      });
+    });
+    describe('with dojoConfig option', function () {
+      var dojoConfig = {
+        async: true,
+        packages: [
+          {
+            location: 'path/to/somelib',
+            name: 'somelib'
+          }
+        ]
+      };
+      beforeAll(function () {
+        spyOn(document.body, 'appendChild').and.callFake(function (el) {
+          // call the onload callback
+          el.onload();
+        });
+        esriLoader.bootstrap(undefined, {
+          dojoConfig: dojoConfig
+        });
+      });
+      it('should have set global dojoConfig', function () {
+        expect(window.dojoConfig).toEqual(dojoConfig);
       });
     });
     describe('when called twice', function () {