+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+- set `window.dojoConfig` by passing as an option to `bootstrap()`
+
 ## 1.2.0
 
 ### Added